32-bit cleanup for embedbin
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@4910 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/kern/embedbin.h
@@ -6,7 +6,7 @@
 struct embed_bin {
     const char *name;
     const uint8_t *buf;
-    uint64_t size;
+    uintptr_t size;
 };
 
 extern struct embed_bin embed_bins[];
